Renamed multiple files to be consistent
