+++ public/controllers/blank-screen-controller.js
