+++ public/controllers/health-check.js
