+++ public/kibana-integrations/kibana-discover.js
