+++ public/kibana-integrations/kibana-filter-bar.js
