+++ public/kibana-integrations/kibana-visualization.js
