+++ public/kibana-integrations/loader/embedded-visualize-handler.js
