+++ public/kibana-integrations/loader/loader-template.html
