+++ public/services/pattern-handler.js
